pyproject.toml
[project]
name = "tbot"
version = "2.0"
requires-python = ">=3.13"
description = "Twitch, Discord and YouTube bot"
dependencies = [
    "twitchAPI==4.3.1",
    "fastapi==0.115.11",
    "sqlalchemy==2.0.38",
    "yaml-settings-pydantic==2.3.2",
    "pydantic[email]==2.10.6",
    "redis==5.2.1",
    "httpx[http2]==0.28.1",
    "httpx_auth==0.23.1",
    "authlib==1.4.0",
    "uvicorn==0.34.0",
    "honcho==2.0.0",
    "arq==0.26.3",
    "asyncmy==0.2.10",
    "alembic==1.14.1",
    "uuid6==2024.7.10",
    "humanize==4.11.0",
    "python-dateutil==2.9.0.post0",
    "sqlakeyset==2.0.1739066250",
    "PyJWT==2.10.1",
    "PyMySQL==1.1.1",
    "more_itertools==10.6.0",
    "async-lru==2.0.5",
    "asyncclick>=8.1.8",
    "pytz>=2025.2",
    "unidecode>=1.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest-httpx==0.35.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
    "pytest-mock==3.14.0",
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core>=3.2,<4"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "I", "FAST", "ANN"]
ignore = ['ANN401']

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
